Fixes dragging folders to move them.
A user expects to move folders by dragging them. Sadly,
this functionality failed to execute with no user feedback.
That is, the user was able to drag the folders, but dropping
them didn’t move the dragged folder into the new parent
folder, even though the move was valid.

In short, this issue rose because the `folderId` passed
around was not the expected value.

To go into more detail:

The first issue was in the renderer process, which passed
the wrong `folderKey` to the main process method
`moveFolder`. Thus, the `moveFolder` method always led
into the `folder.isEmpty()` condition.

The second issue was later in the `moveFolder` method,
which was not updating the `parentFolderId` property to
the correct folder ID.